requirements.txt
appdirs==1.4.3
attrs==19.3.0
beautifulsoup4==4.8.2
black==19.10b0
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cssselect==1.1.0
fake-useragent==0.1.11
idna==2.8
lxml==4.4.2
parse==1.14.0
pathspec==0.7.0
pyee==6.0.0
pyppeteer==0.0.25
pyquery==1.4.1
PySimpleGUI==4.15.2
python-docx==0.8.10
regex==2020.1.8
requests==2.22.0
requests-html==0.10.0
six==1.13.0
soupsieve==1.9.5
toml==0.10.0
tqdm==4.41.1
typed-ast==1.4.1
urllib3==1.25.7
w3lib==1.21.0
websockets==8.1
